requirements.txt
streamlit
ultralytics
opencv-python-headless
Pillow
numpy
pandas
torch
torchvision
gdown
